refactor(vr): audit cleanup for BSPreCulledObjects and FasterWorkshop
BSPreCulledObjects VR:
- Split the all-inline header into a declaration-only .h plus a .cpp,
  matching the rest of the codebase and speeding incremental builds.
- Replace the single shared_mutex over the ID->3D map with 64-way
  sharded locking keyed by formID. Writes stay O(1); a writer now only
  blocks the ~1/64 of readers hitting the same shard, cutting
  contention at cell-transition write bursts. (A full lock-free
  copy-on-write snapshot was rejected: each entry is an NiPointer, so
  copying the map on every write would trigger a refcount storm across
  every live 3D object.)
- Remove dead code: the unused span find(), TrackVisibility,
  CallVisibilityCallbacks, and RemoveVisibilityCallBackForID, dropping
  three version-specific hardcoded offsets along with them.
- Replace QueryPerformanceCounter with std::chrono::steady_clock in the
  perf-counter heartbeat, removing the windows.h dependency.
- Add a static_assert that kBatchSize divides the log-gate period.

The double-write to the stock hash table in UpdateIDto3DMap is kept
deliberately: a Ghidra xref of Get3DForID (0x1427E0980) shows two
referencers we do not hook (a direct call in FUN_1428757B0 and a
function-pointer table entry), both of which read the stock table, so
it must stay populated.

FasterWorkshop VR:
- Guard CheckForValidChildren against self-referential FormLists with a
  depth cap. The cap lives in an internal 3-arg worker; the detour
  entry point keeps the engine's 2-arg ABI so the depth can't be fed
  garbage from r8.
- Trim the Windows define block to the two macros actually needed.

Both patches verified in-game on Fallout4VR 1.2.72: clean install,
full hot paths exercised (BSPreCulled telemetry 0 real misses across
millions of lookups; FasterWorkshop COBJ map built, recipes render),
no crashes.
