+++ src/CMakeLists.txt
@@ -68,6 +68,7 @@ set(SOURCE_FILES
 	"${SOURCE_DIR}/Patches/AchievementsPatch.h"
 	"${SOURCE_DIR}/Patches/BSMTAManagerPatch.h"
 	"${SOURCE_DIR}/Patches/BSPreCulledObjectsPatch.h"
+	"${SOURCE_DIR}/Patches/BSPreCulledObjectsPatchVR.cpp"
 	"${SOURCE_DIR}/Patches/BSPreCulledObjectsPatchVR.h"
 	"${SOURCE_DIR}/Patches/BSTextureStreamerLocalHeapPatch.h"
 	"${SOURCE_DIR}/Patches/FasterWorkshopPatch.cpp"
